Remove clean_spec from criss_cross.
Separated clean_spec so it can become its own independent ciao contrib tool. Added the most recent version of load_hetg_resp so it will be compatible with that sherpa contrib tool on release. Modified .gitignore so these files will be backed up.
