+++ ciao_contrib/criss_cross/.gitignore
@@ -10,3 +10,5 @@
 !README.md
 !constants.py
 !tests
+!clean_spec.py
+!load_hetg_resp.py
