+++ ciao_contrib/criss_cross/criss_cross.py
@@ -1851,553 +1851,7 @@ class TimeLogger:
         return self.total_time
 
 
-def find_resp_files(pha2_file_par, resp_type_par, resp_dir_par=None):
-    """
-    Identifies ARFs and RMFs that are associated with an HETG PHA2 file.  This currently utilizes the 'response' and
-    'tg' directories created by the data systems and chandra_repro pipelines and thus might not work for all PHA2 HETG
-    files.
 
-    Parameters
-    ----------
-
-    pha2_file_par : PHA2 fits file
-         PHA2 spectrum which is typically provided in the chandra archive of a downloaded HETG observation or after
-         running chandra_repro on a chandra HETG obsID.
-    resp_type_par: 'arf' or 'rmf'
-        The type of response files for matchign to PHA spectra
-    resp_dir_par: directory
-        The directory where the HETG response files associated with the PHA2 file are located. If none provided, it
-        will attempt to search for them.
-
-    Returns
-    ----------
-    resp_list_par: list
-    Returns a list of ARF or RMF files found.
-
-    """
-
-    # if user enters 'ARF' or 'RMF' then lower case for later glob use
-    resp_type_par = resp_type_par.lower()
-
-    # check to make sure responses are either arf or rmf
-    if resp_type_par != "arf" and resp_type_par != "rmf":
-        raise ValueError("Response type must be either arf or rmf")
-
-    # load the pha2 file and obtain the crate where relevant information is stored
-    pha2_dataset = read_pha(pha2_file_par)
-    spec_crate_par = pha2_dataset.get_crate(2)
-
-    # identify the number of spectra in the PHA2 file
-    num_spec_par = len(spec_crate_par.TG_M.values)
-
-    # determine which pipeline the PHA2 file came from (archive vs CIAO user)
-    creator_key = get_keyval(spec_crate_par, "CREATOR")
-
-    if "Version DS" in creator_key:
-        # if users pha2_file_par is a long path or a single file, strip the name and check for the root so the resp
-        # glob doesn't get any extra unrelated files in the response dir.
-
-        # Note, the strings in this section are based on the DS standard naming.
-        pha_name = Path(pha2_file_par).name
-        pha_root = pha_name.partition("_pha2.fits")[0]
-
-        if resp_dir_par is not None:  # use provided resp_dir_par
-            resp_list_par = glob.glob(
-                f"{resp_dir_par}/{pha_root}*{resp_type_par}2.fits*"
-            )
-        else:
-            pha_dir = Path(
-                pha2_file_par
-            ).parent  # need to get directory where PHA2 file is located
-            resp_list_par = glob.glob(
-                f"{pha_dir}/responses/{pha_root}*{resp_type_par}2.fits*"
-            )
-
-    elif (
-        "Version CIAO" in creator_key
-    ):  # this is produced via chandra_repro or user custom spectral extraction
-        # read the header to search for the PHA2 root name which is often the same root as the responses
-        hist = get_history_records(spec_crate_par)
-        pha_root = ""  # set to '' for checking later in case pha_root not found
-
-        # if chandra_repro was used to produce the PHA2 file then a ':root" value is saved in the header history.
-        # This identifies that value.
-        for i in range(0, len(hist)):
-            if ":root=" in hist[i][1]:  # find the line where the :root command was used
-                root_line = (
-                    hist[i][1].split("=", 1)[1].strip()
-                )  # strip the unnecessary stuff but leaves the extra spaces
-                pha_root = root_line.split(" ")[
-                    0
-                ]  # remove everything after the root value
-                break  # stops searching hist for the appropriate line after it is found
-
-        # if pha_root is not overwritten at this point then throw error because it means it was not found
-        if pha_root == "":
-            raise ValueError(
-                "Could not identify PHA2 file root. Please load responses manually."
-            )
-
-        # use glob and pha_root to find the responses
-        if resp_dir_par is not None:  # use provided resp_dir_par
-            resp_list_par = glob.glob(f"{resp_dir_par}/{pha_root}*.{resp_type_par}")
-        else:
-            pha_dir = Path(
-                pha2_file_par
-            ).parent  # need to get directory where PHA2 file is located
-            resp_list_par = glob.glob(
-                f"{pha_dir}/tg/{pha_root}*.{resp_type_par}"
-            )  # use the root name of the PHA2 file to ID the responses. This way users can have many extractions in a
-            # single dir but it will only grab the appropriate ones.
-
-    # if the creator of the PHA2 file is not DS or CIAO then exit.
-    else:
-        raise ValueError(
-            "Cannot determine the creator of PHA2 file and responses will have to be loaded manually"
-        )
-
-    # check to make sure at least some files were found
-    if len(resp_list_par) < 1:
-        raise ValueError(
-            "Could not identify responses. Please load responses manually."
-        )
-
-    # check that the length of the arf or RMF lists match the number of PHA spec in the PHA2 file
-    if len(resp_list_par) != num_spec_par:
-        print(
-            f"WARNING-- The identified number of {resp_type_par.upper()}s [{len(resp_list_par)}] does not match the number of PHA spectra [{num_spec_par}] in the PHA2 file. Only the responses found will be included.\n"
-        )
-
-    return resp_list_par
-
-
-def match_resp_order(pha2_file_par, resp_list_par, resp_type_par):
-    """
-    This uses the PHA2 file structure and matches the input response file list/array (resp_list_par) to the appropriate
-    PHA2 spectrum using the header keywords tg_m, tg_part and obsid. This function is designed to take the output of
-    find_resp_files() and put them in order of the PHA2 spectra.
-
-    Parameters
-    ----------
-
-    pha2_file_par : PHA2 fits file
-         PHA2 spectrum which is typically provided in the chandra archive of a downloaded HETG observation or after
-         running chandra_repro on a chandra HETG obsID.
-    resp_type_par: 'arf' or 'rmf'
-        The type of response files for matchign to PHA spectra
-    resp_dir_par: directory
-        The directory where the HETG response files associated with the PHA2 file are located. If none provided, it
-        will attempt to search for them.
-
-    Returns
-    ----------
-
-    matched_resp_list_par: array
-        Returns an array of ARF or RMF file paths matched to the order (arrangement) of the input PHA2 file. The string
-        value 'no match' is returned for elements where there is a spectrum in the PHA2 file but no matching response
-        file.
-
-    """
-
-    # check to make sure responses are either arf or rmf
-    if resp_type_par != "arf" and resp_type_par != "rmf":
-        raise ValueError("Response type must be either arf or rmf")
-
-    # reads in the pha2 file and checks how many spectra (orders) it includes.
-    pha2_dataset = read_pha(pha2_file_par)
-    spec_crate_par = pha2_dataset.get_crate(2)
-
-    # identify the file arrangement of the HETG orders and HEG/MEG arms via the PHA2 file
-    tg_m_arr = spec_crate_par.TG_M.values
-    tg_part_arr = spec_crate_par.TG_PART.values
-    tg_obsid = get_keyval(spec_crate_par, "OBS_ID")
-
-    # determine the number of spectra in the pha2 file based on the length of one of the tg arrays:
-    num_spec_pha2 = len(tg_m_arr)
-
-    # warn user if the number of PHA2 spectra are different than the number of response files (either found
-    # automatically or provided manually)
-    if len(resp_list_par) != num_spec_pha2:
-        print(
-            f"\nWARNING -- The number of {resp_type_par} files [{len(resp_list_par)}] does not equal the number of spectra in the PHA2 file [{num_spec_pha2}]. Only responses that match with a spectrum will be included.\n"
-        )
-
-    # create empty lists to later append HEG/MEG arm, order and obsID values from the response files headers
-    resp_m_arr = []
-    resp_tg_part_arr = []
-    resp_obsid_arr = []
-
-    # read each response file and append appropriate header value
-    for i in range(0, len(resp_list_par)):
-        resp_data = read_file(resp_list_par[i])
-        resp_m_arr.append(get_keyval(resp_data, "TG_M"))
-        resp_tg_part_arr.append(get_keyval(resp_data, "TG_PART"))
-        resp_obsid_arr.append(get_keyval(resp_data, "OBS_ID"))
-
-    # convert obsid lists to numpy arrays for later use with np.where() for obsID checking
-    tg_obsid = np.array(tg_obsid)
-    resp_obsid_arr = np.array(resp_obsid_arr)
-
-    # create an empty object array the same size as the PHA2 file (number of spectra) to hold either 'no match' or the
-    # path to the matched response file
-    matched_resp_list_par = np.array([""] * num_spec_pha2, dtype="object")
-
-    # for each spectra, use tg_m, tg_part and obsID to match to a single response file. Print appropriate errors.
-    for i in range(0, num_spec_pha2):
-        match_resp = np.where(
-            (resp_m_arr == tg_m_arr[i])
-            & (resp_tg_part_arr == tg_part_arr[i])
-            & (resp_obsid_arr == tg_obsid)
-        )[0].tolist()
-        if len(match_resp) == 1:
-            matched_resp_list_par[i] = resp_list_par[
-                match_resp[0]
-            ]  # this is the element in the response array that match the i_th element of the PHA2 file.
-        elif len(match_resp) > 1:
-            raise ValueError(
-                f"ERROR, more than one {resp_type_par.upper()} match found for TG_M={tg_m_arr[i]}, TG_PART={tg_part_arr[i]} and obsID={tg_obsid}. "
-            )
-        elif len(match_resp) == 0:
-            matched_resp_list_par[i] = "no match"
-            print(
-                f"Warning, no {resp_type_par.upper()} found for TG_M={tg_m_arr[i]}, TG_PART={tg_part_arr[i]} and obsID={tg_obsid}."
-            )
-        else:
-            raise ValueError(
-                f"ERROR - Something with wrong identifying {resp_type_par.upper()}s for TG_M={tg_m_arr[i]}, TG_PART={tg_part_arr[i]} and obsID={tg_obsid}"
-            )
-
-    # report the files matched to the screen in a nice format so it is clear it worked or didn't work
-    print("\nThe following response files were found\n")
-
-    print()
-    for i in range(0, num_spec_pha2):
-        print(
-            f"{tg_part_name[tg_part_arr[i]]}{tg_m_arr[i]:+n} -- {resp_type_par.upper()}: {matched_resp_list_par[i]}"
-        )
-    print()
-
-    # returns the array of matched responses in the same order as the PHA2 spectra
-    return matched_resp_list_par
-
-
-def clean_spec(cc_table, pha_file, spec_root, arf_file=None, resp_dir=None):
-    """
-    Uses confusion tables produced by CrissCross to create 'cleaned' PHA1 or PHA2 spectra and ARF response files. The
-    confusion tables identify portions of a source's spectrum that may have erroneous events that should not be
-    included in any spectral analysis. After identifying the PHA type (1 vs 2), this function checks to make sure the
-    ARF files provided match the PHA spectra (order and arm) and uses clean_data() to set the appropriate PHA
-    (COUNTS, STAT_ERROR) and ARF (SPECRESP, FRACEXPO) columns to zero where confusion occurs. This function writes a
-    'cleaned' copy of the input PHA and ARF files.
-
-    Parameters
-    ----------
-
-    cc_table: fits table
-        CrissCross produced confusion table for a single source and single obsID.
-    pha_file: fits file
-        HETG PHA1 or PHA2 spectral file of the source that needs cleaning.
-    spec_root: string
-        A root for file naming purposes.
-    arf_file: string or list
-        A file path or list of file paths to ARFs matching the input pha file.
-    resp_dir: directory
-        The directory where the ARFs associated with the pha_file is stored.
-
-    """
-
-    def clean_data(
-        cc_table,
-        pha_crate,
-        arf_file_var,
-        pha_element,
-        conf_flag_var="confused",
-    ):
-        """
-        Creates copies of the relevant PHA1/PHA2 and ARF columns and modifies them (sets to zero) using a CrissCross
-        confusion table.
-
-        Parameters
-        ----------
-
-        cc_table: fits table
-            CrissCross produced confusion table for a single source and single obsID.
-        pha_crate: Crate object
-            The Crate data for a single spectrum (e.g., HEG+1)
-        arf_file_var: string
-            A file path to an ARF matching the input pha file.
-        pha_element: integer
-            The element of the PHA2 file assocaited with the spectrum 'pha_crate'. If standard HETG PHA2 file
-            [order,element] = HEG: -3,0; -2,1; -1,2; +1,3; +2,4; +3,5 MEG: -3,6; -2,7; -1,8; +1,9; +2,10; +3,11)
-        conf_flag_var: string 'confused'
-            The string associated with spectral confusion. If 'confused', spectra will be cleaned by setting all
-            'confused' values to zero. A future update will allow 'confused' and/or 'warn' to be zeroed out.
-
-
-        Returns
-        ----------
-
-        cleaned_spec_var, cleaned_staterr_var, cleaned_specresp_var, cleaned_fracexpo_var: arrays
-            Returns a copy of the SPEC, STAT_ERR, SPECRESP, and FRACEXPO arrays with values associated with the
-            identified wavelengths of confusion set to zero (or 1.86603 for STAT_ERR).
-        """
-
-        # reads in the confusion and PHA data
-        cc_data = read_file(cc_table)
-        pha_data_var = pha_crate.get_crate(2)
-        arf_data_var = read_file(arf_file_var)
-
-        # identifies ARF tg_m, tg_order, and tg_part 
-        arf_tg_part = get_keyval(arf_data, "TG_PART")
-        arf_tg_m = get_keyval(arf_data, "TG_M")
-
-        # PHA1 and PHA2 files have to be treated slightly differently because of how crates stores values. This is to
-        # avoid having to slice off each crate spectrum from the PHA2 file.
-        if is_pha_type1(pha_crate):
-            counts_arr = pha_data_var.COUNTS.values
-            stat_err_arr = pha_data_var.STAT_ERR.values
-            specresp_arr = arf_data_var.SPECRESP.values
-            fracexpo_arr = arf_data_var.FRACEXPO.values
-            bin_low_arr = pha_data_var.BIN_LO.values
-            bin_high_arr = pha_data_var.BIN_HI.values
-
-        elif is_pha_type2(pha_crate):
-            counts_arr = pha_data_var.COUNTS.values[pha_element]
-            stat_err_arr = pha_data_var.STAT_ERR.values[pha_element]
-            specresp_arr = (
-                arf_data_var.SPECRESP.values
-            )  # ARFs are not in PHA2 (array) format and thus don't need a pha_element
-            fracexpo_arr = (
-                arf_data_var.FRACEXPO.values
-            )  # ARFs are not in PHA2 (array) format and thus don't need a pha_element
-            bin_low_arr = pha_data_var.BIN_LO.values[pha_element]
-            bin_high_arr = pha_data_var.BIN_HI.values[pha_element]
-
-        else:
-            raise ValueError("PHA datatype must be 1 or 2")
-
-        # copies the counts and stat_err column of the PHA file for modification
-        cleaned_spec_var = counts_arr.copy()
-        cleaned_staterr_var = stat_err_arr.copy()
-
-        # copies the SPECRESP and fracexpo columns from the arf
-        cleaned_specresp_var = specresp_arr.copy()
-        cleaned_fracexpo_var = fracexpo_arr.copy()
-
-        # for every row of the confusion table that match the input PHA spectrum order and tg_part, identify the
-        # elements (rows) associated with wavelengths (bin_low and bin_hi) that need to be cleaned. Note, this assumes
-        # the PHA bin_lo and bin_hi values are identical to the ARF file (which should be the case).
-        for i in range(0, len(cc_data.wave_low.values)):
-            if (
-                cc_data.flag.values[i] == conf_flag_var
-                and cc_data.grating_type.values[i] == tg_part_name[arf_tg_part]
-                and cc_data.order.values[i] == arf_tg_m
-            ):
-                elements_to_clean = np.where(
-                    (bin_low_arr >= cc_data.wave_low.values[i])
-                    & (bin_high_arr <= cc_data.wave_high.values[i])
-                )  # identify elements
-
-                # clean PHA (spectrum)
-                cleaned_spec_var[elements_to_clean] = (
-                    0.0  # set elements that overlap to zero
-                )
-                cleaned_staterr_var[elements_to_clean] = (
-                    1.86603  # double check that this makes sense and the stat_err is always this value for zero counts.
-                    # I suspect instead I should take min of this column and set it to that.
-                )
-
-                # clean ARF (response)
-                cleaned_specresp_var[elements_to_clean] = 0.0
-                cleaned_fracexpo_var[elements_to_clean] = 0.0
-
-        # return the cleaned arrays values.
-        return (
-            cleaned_spec_var,
-            cleaned_staterr_var,
-            cleaned_specresp_var,
-            cleaned_fracexpo_var,
-        )
-
-    # Read the PHA file and determine if PHA1 or PHA2. Uses read_pha() to bring along the necessary PHA extensions.
-    pha_crate_dataset = read_pha(pha_file)
-
-    if is_pha_type1(pha_crate_dataset):
-        pha_data = pha_crate_dataset.get_crate(2)  # Crate 2 contains the PHA data
-        arf_data = read_file(arf_file)
-
-        # determine the heg/meg arm, order and obsID of the PHA spectrum and make sure it matches the ARF
-        tg_part = get_keyval(
-            pha_data, "TG_PART"
-        )  # tg_part = 1 = heg; tg_part = 2 = meg
-        tg_m = get_keyval(pha_data, "TG_M")
-        tg_obs = get_keyval(pha_data, "OBS_ID")
-        tg_part_arf = get_keyval(arf_data, "TG_PART")
-        tg_m_arf = get_keyval(arf_data, "TG_M")
-        tg_obs_arf = get_keyval(arf_data, "OBS_ID")
-
-        # if parameters dont match between ARF and PHA then throw error
-        if tg_obs != tg_obs_arf or tg_m != tg_m_arf or tg_part != tg_part_arf:
-            raise ValueError(
-                "One of the following is not consistent between the PHA file and ARF: HEG/MEG arm, order, obsID"
-            )
-
-        # use clean_data() to create copies of the appropriate PHA and ARF arrays where wavelenghts with confusion are
-        # set to 0
-        cleaned_spec, cleaned_staterr, cleaned_specresp, cleaned_fracexpo = clean_data(
-            cc_table=cc_table,
-            pha_crate=pha_crate_dataset,
-            arf_file_var=arf_file,
-            pha_element=0,
-            conf_flag_var="confused",
-        )
-
-        # replaces the original arrays in the loaded crates with the new arrays
-        pha_data.COUNTS.values = cleaned_spec
-        pha_data.STAT_ERR.values = cleaned_staterr
-        arf_data.SPECRESP.values = cleaned_specresp
-        arf_data.FRACEXPO.values = cleaned_fracexpo
-
-        # appends the original files to the history for record keeping
-        pha_data.add_record(
-            "HISTORY",
-            f"This cleaned spectrum was created using the PHA file: {pha_file} and the CrissCross cleaning table: {cc_table}. ",
-        )
-        arf_data.add_record(
-            "HISTORY",
-            f"This cleaned ARF was created using the ARF file: {arf_file} and the CrissCross cleaning table: {cc_table}. ",
-        )
-
-        update_crate_checksum(pha_data)
-        update_crate_checksum(arf_data)
-
-        # setup tgpart and order to be consistent with values in confusion tables for file naming.
-        pha_arm = tg_part_name[tg_part]
-        pha_order = f"{tg_m:+n}"
-
-        # saves a new file while maintaining the original header.
-        write_pha(
-            pha_crate_dataset,
-            f"{spec_root}_obsid_{tg_obs}_{pha_arm}{pha_order}_cleaned.pha",
-            clobber=True,
-        )
-        write_file(
-            arf_data,
-            f"{spec_root}_obsid_{tg_obs}_{pha_arm}{pha_order}_cleaned.arf",
-            clobber=True,
-        )
-
-    # PHA2 files need to be treated a little different because they are arrays of arrays and order/arm info is not in
-    # header. Also the arfs may be out of order if user provides a list of arfs.
-    elif is_pha_type2(pha_crate_dataset):
-        # if user enters their own arf or list of arfs then make sure they match the PHA2 file format (e.g., pha2 meg+1
-        # has to be same array element as arf meg+1)
-        if arf_file is not None:
-            if type(arf_file) is str:
-                arf_file = list(
-                    [arf_file]
-                )  # make sure arf_file variable is a list just in case user enters a single file as 'file1' instead of ['file1'].
-
-            # check and arrange that the user input arf file(s) are in the correct order of the PHA2 spectra.
-            # matched_resp_list will create an array that matches the PHA2 format.
-            matched_resp_list = match_resp_order(
-                pha2_file_par=pha_file, resp_list_par=arf_file, resp_type_par="arf"
-            )
-
-        # if user doesn't enter arfs then try to find them either using the user included response dir or the standard
-        # CIAO-produced file structure
-        else:
-            print(
-                "Warning, no ARF response files provided in parameter arf_file. Attempting to find them."
-            )
-            resp_list = find_resp_files(
-                pha2_file_par=pha_file, resp_type_par="arf", resp_dir_par=resp_dir
-            )
-
-            if len(resp_list) == 0:
-                raise ValueError(
-                    "No response files found. Try including a directory with resp_dir_par or include a list of response paths with arf_file parameter."
-                )
-
-            # matched_resp_list will create an array that matches the PHA2 format.
-            matched_resp_list = match_resp_order(
-                pha2_file_par=pha_file, resp_list_par=resp_list, resp_type_par="arf"
-            )
-
-        # read the PHA data and obtain the tg_m, tg_part and obsID values
-        pha_data_full = pha_crate_dataset.get_crate(2)
-
-        tg_m_arr = pha_data_full.TG_M.values
-        tg_part_arr = pha_data_full.TG_PART.values
-        tg_obs = get_keyval(pha_data_full, "OBS_ID")
-
-        pha_order_arr = [
-            f"{i:+n}" for i in tg_m_arr
-        ]  # this is to make sure the order has a + or - sign in front of it for file naming purposes
-        pha_arm_arr = [
-            tg_part_name[i] for i in tg_part_arr
-        ]  # this is to make sure the arm is named 'heg' or 'meg' for file naming purposes
-
-        # only run clean_data() for the spectra that have a matched response file so identify which elements of the
-        # matched_resp_list have both a spectrum and associated ARF.
-        spec_to_clean = np.where(
-            matched_resp_list != "no match"
-        )[
-            0
-        ]  # note, the 'no match' string comes from match_resp_order() so be careful if that changes.
-
-        # run clean_data() for every arm and order which have a spectrum and matching ARF
-        for i in spec_to_clean:
-            arf_data = read_file(
-                matched_resp_list[i]
-            )  # load the arf from match_resp_list because it is already matched to the appropriate PHA2 file
-
-            cleaned_spec, cleaned_staterr, cleaned_specresp, cleaned_fracexpo = (
-                clean_data(
-                    cc_table=cc_table,
-                    pha_crate=pha_crate_dataset,
-                    arf_file_var=matched_resp_list[i],
-                    pha_element=i,
-                    conf_flag_var="confused",
-                )
-            )
-
-            # replaces the original arrays with the new arrays
-            pha_data_full.COUNTS.values[i] = cleaned_spec
-            pha_data_full.STAT_ERR.values[i] = cleaned_staterr
-            arf_data.SPECRESP.values = cleaned_specresp
-            arf_data.FRACEXPO.values = cleaned_fracexpo
-
-            # update the ARF header and write out the file
-            arf_data.add_record(
-                "HISTORY",
-                f"This cleaned ARF was created using the ARF file: {matched_resp_list[i]} and the CrissCross cleaning table: {cc_table}. ",
-            )
-            update_crate_checksum(arf_data)
-
-            write_file(
-                arf_data,
-                f"{spec_root}_obsid_{tg_obs}_{pha_arm_arr[i]}{pha_order_arr[i]}_cleaned.arf",
-                clobber=True,
-            )
-
-        # appends the original files to the history of the new file for record keeping
-        pha_data_full.add_record(
-            "HISTORY",
-            f"This cleaned spectrum was created using the PHA file: {pha_file} and the CrissCross cleaning table: {cc_table}. ",
-        )
-
-        update_crate_checksum(pha_data_full)
-
-        # saves a new PHA2 file while maintaining the original header.
-        write_pha(
-            pha_crate_dataset, f"{spec_root}_obsid_{tg_obs}_cleaned.pha2", clobber=True
-        )
-
-    else:
-        raise ValueError("Input PHA file was not a PHA1 or PHA2 type file")
-
-    return ()
 
 
 ######### MAIN CrissCross RUN FUNCTION ##############
